Move modules around, use Read instance for format
