--- neuron/src/lib/Neuron/Zettelkasten/Zettel/Format.hs
@@ -1,32 +1,0 @@
-{-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE LambdaCase #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE NoImplicitPrelude #-}
-
-module Neuron.Zettelkasten.Zettel.Format where
-
-import Data.Aeson (FromJSON, ToJSON)
-import Relude
-
-data ZettelFormat
-  = ZettelFormat_Markdown
-  | ZettelFormat_Org
-  deriving (Eq, Ord, Show, Generic, FromJSON, ToJSON)
-
-zettelFormatToExtension :: ZettelFormat -> Text
-zettelFormatToExtension = \case
-  ZettelFormat_Markdown -> ".md"
-  ZettelFormat_Org -> ".org"
-
-extensionToZettelFormat :: Text -> Maybe ZettelFormat
-extensionToZettelFormat = \case
-  ".md" -> Just ZettelFormat_Markdown
-  ".org" -> Just ZettelFormat_Org
-  _ -> Nothing
-
-formatDescToZettelFormat :: Text -> Maybe ZettelFormat
-formatDescToZettelFormat = \case
-  "markdown" -> Just ZettelFormat_Markdown
-  "org" -> Just ZettelFormat_Org
-  _ -> Nothing
