+++ neuron/test/Neuron/Zettelkasten/ZettelSpec.hs
@@ -9,10 +9,10 @@ where
 
 import Data.TagTree
 import Data.Time.Calendar
-import Neuron.Markdown
+import Neuron.Reader.Markdown
+import Neuron.Reader.Type
 import Neuron.Zettelkasten.ID
 import Neuron.Zettelkasten.Zettel
-import Neuron.Zettelkasten.Zettel.Format
 import Neuron.Zettelkasten.Zettel.Meta (Meta)
 import Relude
 import Test.Hspec
